Added a first draft of the staging concept.
This is not yet coined to wikidata, but in highly experimental state.

Abstract: A stage is a set of tasks, that are to be executed in a
certain order i.e. download files from newest to oldest.

Once a stage has output available to be processed further by another
stage the output can be forwarded via synchronized queues the the next
stage(s). The stages itself run in seperate threads and are modelled as
producer-consumer problems.

In WDTK the stages might be (referring to the DumpProcessingExample):
1) find/download files
2) extract and filter revisions
3) parse revisions into JSONObjects
4) measure statistics