The Stage class now is more uniform and streamlined in respect to handling producers and consumers.
As a result, the StageManager has undergone a diet and became less
all-mighty and more focused to the task it should achieve (relieving the
user of doing to much boring setup and track-keeping)

Also the stages now can inform each other of new input and terminate
themselves if there are no more producers available and all input is
processed. This also is to relieve the user of management duty.

The StageManager now has its own thread, so the user can do other stuff
while it and the managed stages are running. I hope this is especially
useful for people who want to put an UI atop the WDTK or run other code
in the main thread in parallel.

The examples have been changed accordingly.